server/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
blinker==1.8.2
    # via flask
click==8.1.7
    # via flask
flask==3.0.3
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via flask
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
packaging==24.1
    # via gunicorn
pillow==10.4.0
    # via -r requirements.in
str2bool==1.1
    # via -r requirements.in
werkzeug==3.0.4
    # via flask
